backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
firebase-admin
langchain
langchain-google-genai
python-jose[cryptography]
passlib[bcrypt]